+++ order/src/main/java/store/_0982/order/application/settlement/SellerSettlementProcessor.java
@@ -3,8 +3,8 @@ package store._0982.order.application.settlement;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import store._0982.order.infrastructure.client.ProductFeignClient;
-import store._0982.order.infrastructure.client.dto.GroupPurchaseInternalInfo;
+import store._0982.order.client.ProductFeignClient;
+import store._0982.order.client.dto.GroupPurchaseInternalInfo;
 
 import java.util.List;
 import java.util.UUID;
