+++ order/src/main/java/store/_0982/order/client/dto/GroupPurchaseInternalInfo.java
@@ -1,4 +1,4 @@
-package store._0982.order.infrastructure.client.dto;
+package store._0982.order.client.dto;
 
 import java.util.UUID;
